extract csv helpers to lib/csv
Move escapeCsvValue and downloadCsv into a shared module so the same
fixes (Firefox detached-anchor click, Excel UTF-8 BOM) only need to live
in one place. Both user-export pages now serialize their rows locally
and hand the headers, rows, and filename to the shared helper.
